// CAC/src/utils/models/medicion/medicion.ts
export class objMedicion{
    id: number = 0;
    idCobertura: number = 0;
    idPeriodo: number = 1;
    descripcion: string = "";
    activo: boolean = true;
    createdBy: string = "";
    createdDate: string = "";
    modifyBy: string = "";
    modifyDate: string = "";
    estado: number = 28;
    fechaInicioAuditoria: string = "";
    fechaFinAuditoria: string = "";
    fechaCorteAuditoria: string = "";
    lider: string = "";
    resolucion: string = "";
    nombre: string = "";    
    status: number = 0;
    
    
    
    
    
    
    
    
    
    
    
}